Re-activate two more tests under the refactored code.
git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@700 d9297710-12f9-482b-b11f-fb677dc88bbb

+++ pditi/mesh_input/quick_grid/QuickGrid.cxx
@@ -955,7 +955,7 @@ QuickGridData TensorProductSolidCylinder::allocatePdGridData() const {
 
 }
 
-std::pair<Cell3D,QuickGridData> TensorProductSolidCylinder::computePdGridData(int proc, Cell3D cellLocator, QuickGridData& pdGridData, NormFunctionPointer norm) const {
+std::pair<Cell3D,QuickGridData> TensorProductSolidCylinder::computePdGridData(size_t proc, Cell3D cellLocator, QuickGridData& pdGridData, NormFunctionPointer norm) const {
 
 	/**
 	 * This routine attempts to make as close of an analogy to CellsPerProcessor3D::computePdGridData(...)
